Split heat_transfer into multiple files
