+++ pygfunction/heat_transfer/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from .finite_line_source import finite_line_source, finite_line_source_equivalent_boreholes_vectorized
+from .finite_line_source_vertical import finite_line_source_vertical
+from .finite_line_source_inclined import finite_line_source_inclined
